refactor: split imposter/core.rs into a core/ module
Pure code-movement, no logic change. The 59KB single impl Imposter block is
split across concern-grouped files, each an additional impl Imposter block that
accesses private state via descendant-module visibility (use super::*):
- core/mod.rs: consts, imports, StubState, the Imposter struct, construction /
  flow-store setup, and the inline test module
- core/matching.rs: request matching, flow-id / scenario-state resolution, form parsing
- core/responses.rs: debug/preview inspection and response-mode dispatch helpers
- core/proxy.rs: proxy record/replay, predicate generation, stub insertion
- core/recording.rs: recorded-request storage and counters
- core/lifecycle.rs: stub add/replace/delete/query and enabled state

parse_form_data and generate_predicates_from_request widen to pub(crate) so the
test module (a sibling of the new submodules) can still reach them. Public paths
(core::Imposter, core::StubState) are unchanged.
